ml-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
python-dotenv==1.0.1
chromadb==0.5.15
posthog<6
google-genai>=1.0.0
numpy==2.1.3
kagglehub==0.3.3

